requirements.txt
python-telegram-bot==20.7
easyocr==1.7.1
python-dotenv==1.0.1
requests==2.31.0

# Google Drive API dependencies
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
